# Faster hflip with OpenCV