@creafz 2019-03-03
@vfdev-5 Could you also test a case with a np.float32 input? On my machine, OpenCV is slower than NumPy:
```
x = np.random.rand(281, 500, 3).astype(np.float32)
```

```
%%timeit -r 7 -n 10
for _ in range(1000):
    hflip(x)
> 509 ms ± 1.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
```
%%timeit -r 7 -n 10
for _ in range(1000):
    hflip_cv2(x)
> 1.62 s ± 329 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```