@vfdev-5 2019-03-03
@creafz yes, I confirm that opencv is slower on f32 similar to your results (not x3 but ~1x). 
Do you have use-cases of images in float32 after reading from disk ? But you're right it make sense to check the performances on int16, int32 input data.