@creafz 2019-03-03
We try to make all transformations compatible with a `np.float32` input, so they can work not only with a `np.uint8` input but with other data types as well. [Here is an example notebook](https://github.com/albu/albumentations/blob/master/notebooks/example_16_bit_tiff.ipynb) that shows how to work with 16-bit tiff images by converting them to float32 using `ToFloat` transformation (and since we can use `ToFloat` to convert any datatype to float32 we don't need to support all the datatypes in our transformations, we just need to support `np.uint8` and `np.float32` inputs).