@ternaus 2019-03-03
@vfdev-5 If you will modify PR so that the implementation of the transform `numpy` vs `cv2` was chosen based on the type of the data it would be really cool.

HorizontalFlip is the most widely used transform and the fact that we are lacking in the performance is not good. Hence, your PR is highly appreciated.