@vfdev-5 2019-03-03
@ternaus I did not have time to write the message to ask whether it is something like that you would like to have and you already merge it :)

Anyway it would be cool to see the benchmark on HorizontalFlip with this add